Review updates and upgrade to latest Hazelcast version
