+++ .gitignore
@@ -15,5 +15,4 @@ target/
 *.dat
 atlassian-ide-plugin.xml
 *dependency-reduced-pom.xml
-hazelcast/src/main/java/com/hazelcast/instance/GeneratedBuildProperties.java
 diagnostics-*.log
