+++ src/main/java/com/hazelcast/commandline/member/MemberCommandLine.java
@@ -5,7 +5,17 @@ import com.hazelcast.commandline.member.names.MobyNames;
 import com.hazelcast.core.HazelcastException;
 import picocli.CommandLine;
 
-import java.io.*;
+import java.io.PrintStream;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.FileInputStream;
+import java.io.ObjectInputStream;
+import java.io.BufferedReader;
+import java.io.FileOutputStream;
+import java.io.InputStreamReader;
+import java.io.ObjectOutputStream;
 import java.lang.reflect.Field;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
